Some fixes

+++ super_scheduler/super_scheduler/utils/schedule/del_schedule.py
@@ -7,10 +7,10 @@ from core.celeryapp import app
 from plugins.super_scheduler.schedule import SCHEDULES, schedule_name2class
 from plugins.super_scheduler.utils.kwargs_parser import KwargsParser, BaseFormat as BaseTaskFormat
 from plugins.super_scheduler.utils.schedule.get_schedule import get_all_schedules_subclasses
+from plugins.super_scheduler.utils.schedule.add_schedule import AddSchedule
 
 
 class ScheduleDeleteFormat(BaseTaskFormat):
-    schedule_subclass: any
 
     @validator('name', allow_reuse=True)
     def name_validator(cls, value: str) -> str:
@@ -18,31 +18,32 @@ class ScheduleDeleteFormat(BaseTaskFormat):
             raise ValueError(f"Not exist schedule with name: {value}")
         return value
 
-    @validator('schedule_subclass')
-    def schedule_subclass_exist(cls, value):
-        if value not in get_all_schedules_subclasses():
-            raise ValueError(f"Not exist schedule with schedule_subclass: {value}")
-        return value
-
 
 class DelSchedule(KwargsParser):
 
     @classmethod
-    def _get_schedule(cls, schedule_kwargs: dict):
+    def delete_by_schedule_subclass(cls, schedule_name: str, schedule_subclass) -> Tuple[bool, Optional[str]]:
         """
-        Get schedule by dict with keys 'name' and 'schedule_subclass'.
+        Delete schedule by schedule subclass.
 
-        :param schedule_kwargs: dict with keys 'name' and 'schedule_subclass'
-        :return: schedule class (django model)
+        :param schedule_name:
+        :param schedule_subclass:
+        :return:
         """
-        schedule_name = schedule_kwargs['name']
-        schedule_subclass = schedule_kwargs['schedule_subclass']
-        return schedule_name2class(schedule_name).from_schedule(schedule_subclass)
+
+        if schedule_name not in SCHEDULES:
+            return False, f"Not exist schedule with name: {schedule_name}"
+
+        if schedule_subclass not in get_all_schedules_subclasses():
+            return False, f"Not exist schedule with schedule_subclass: {schedule_subclass}"
+
+        schedule_name2class(schedule_name).from_schedule(schedule_subclass).delete()
+        return True, None
 
     @classmethod
     def delete(cls, schedule_kwargs: dict) -> Tuple[bool, Optional[str]]:
         """
-        Delete schedule.
+        Delete schedule by schedule kwargs.
 
         :param schedule_kwargs:
         :return:
@@ -52,7 +53,10 @@ class DelSchedule(KwargsParser):
         if schedule_kwargs is None:
             return False, msg
 
-        schedule = cls._get_schedule(schedule_kwargs)
+        schedule, msg = AddSchedule.create(schedule_kwargs)
+        if schedule is None:
+            return False, msg
+
         schedule.delete()
         return True, None
